refactor(admin): colocate product registration components
