+++ src/app/(admin)/admin/products/new/_components/ProductRegistrationForm.quantityImages.test.tsx
