--- src/app/(admin)/admin/products/new/_components/index.tsx
@@ -1,1 +1,0 @@
-export { ProductRegistrationForm } from "./ProductRegistrationForm";
