+++ src/app/(admin)/admin/products/new/_containers/index.tsx
@@ -1,0 +1,1 @@
+export { ProductRegistrationForm } from "./ProductRegistrationForm";
